requirements.txt
django==3.2.4
python-decouple
Pillow
djangorestframework
markdown      
django-filter
drf-yasg[validation] 
djangorestframework-simplejwt